Sources/requirements.txt
certifi==2020.6.20
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imutils==0.5.3
kiwisolver==1.2.0
matplotlib==3.3.2
minimalmodbus==1.0.2
mysql-connector==2.2.9
networkx==2.5
numpy==1.19.3
opencv-python==4.4.0.44
Pillow==8.0.1
pydispatch==1.1.0
pyftpdlib==1.5.6
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyserial==3.5
PySide2==5.15.1
python-dateutil==2.8.1
python-dispatch==0.1.31
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
shiboken2==5.15.1
six==1.15.0
tifffile==2020.10.1
